Use a better way of obtaining the VPN files

+++ xos/synchronizers/vpn/steps/sync_tenantprivilege.py
@@ -3,7 +3,7 @@ import sys
 
 from core.models import TenantPrivilege
 from services.vpn.models import VPN_KIND, VPNService, VPNTenant
-from synchronizers.base.syncstep import SyncStep
+from synchronizers.base.syncstep import DeferredException, SyncStep
 
 parentdir = os.path.join(os.path.dirname(__file__), "..")
 sys.path.insert(0, parentdir)
@@ -22,28 +22,38 @@ class SyncTenantPrivilege(SyncStep):
         return privs
 
     def sync_record(self, record):
+        if (not record.tenant.id):
+            raise DeferredException("Privilege waiting on VPN Tenant ID")
         certificate = self.get_certificate_name(record)
         tenant = VPNTenant.get_tenant_objects().filter(pk=record.tenant.id)[0]
+        if (not tenant):
+            raise DeferredException("Privilege waiting on VPN Tenant")
         # Only add a certificate if ones does not yet exist
-        if (not os.path.isfile(tenant.pki_dir + "/issued/" + certificate + ".crt")):
+        pki_dir = VPNService.get_pki_dir(tenant)
+        if (not os.path.isfile(pki_dir + "/issued/" + certificate + ".crt")):
             VPNService.execute_easyrsa_command(
-                tenant.pki_dir, "build-client-full " + certificate + " nopass")
+                pki_dir, "build-client-full " + certificate + " nopass")
             tenant.save()
         record.save()
 
     def delete_record(self, record):
+        if (not record.tenant.id):
+            return
         certificate = self.get_certificate_name(record)
         tenant = VPNTenant.get_tenant_objects().filter(pk=record.tenant.id)[0]
+        if (not tenant):
+            return
         # If the client has already been reovked don't do it again
-        if (os.path.isfile(tenant.pki_dir + "/issued/" + certificate + ".crt")):
+        pki_dir = VPNService.get_pki_dir(tenant)
+        if (os.path.isfile(pki_dir + "/issued/" + certificate + ".crt")):
             VPNService.execute_easyrsa_command(
-                tenant.pki_dir, "revoke " + certificate)
+                pki_dir, "revoke " + certificate)
             # Revoking a client cert does not delete any of the files
             # to make sure that we can add this user again we need to
             # delete all of the files created by easyrsa
-            os.remove(tenant.pki_dir + "/issued/" + certificate + ".crt")
-            os.remove(tenant.pki_dir + "/private/" + certificate + ".key")
-            os.remove(tenant.pki_dir + "/reqs/" + certificate + ".req")
+            os.remove(pki_dir + "/issued/" + certificate + ".crt")
+            os.remove(pki_dir + "/private/" + certificate + ".key")
+            os.remove(pki_dir + "/reqs/" + certificate + ".req")
             tenant.save()
 
         record.delete()
